fix: bad path separator processing
