+++ vrc-get-gui/app/projects/page.tsx
@@ -55,6 +55,7 @@ import {toastError, toastSuccess, toastThrownError} from "@/lib/toast";
 import {useRemoveProjectModal} from "@/lib/remove-project";
 import {tc, tt} from "@/lib/i18n";
 import {useFilePickerFunction} from "@/lib/use-file-picker-dialog";
+import {pathSeparator} from "@/lib/os";
 
 export default function Page() {
 	const result = useQuery({
@@ -654,7 +655,7 @@ function CreateProject(
 						<Button className={"px-4"} onClick={selectProjectDefaultFolder}>{tc("select directory")}</Button>
 					</div>
 					<Typography variant={"small"} className={"whitespace-normal"}>
-						{tc("the new project will be at <code>{{path}}</code>", {path: `${projectLocation}/${projectName}`})}
+						{tc("the new project will be at <code>{{path}}</code>", {path: `${projectLocation}${pathSeparator()}${projectName}`})}
 					</Typography>
 					<Typography variant={"small"} className={`whitespace-normal ${projectNameStateClass}`}>
 						{projectNameCheck}
